refactor: extract shared types to calendarPageTypes.ts and fix className duplication
Move ImageProps, ImageEditProps, CalendarStyleProps to a dedicated types
file so ImageArea no longer depends on CalendarPage for type imports.
Also fix duplicate gap classes (gap-2 md:gap-2 gap-3 → gap-3 md:gap-2).
